+++ tests/treemodel/batch.js
@@ -6,8 +6,12 @@
 /* bender-tags: treemodel, delta */
 
 'use strict';
+
+import deltas from '/ckeditor5/core/treemodel/delta/basic-deltas.js';
+/*jshint unused: false*/
+
 import Batch from '/ckeditor5/core/treemodel/batch.js';
-import { register } from '/ckeditor5/core/treemodel/batch-base.js';
+import { register } from '/ckeditor5/core/treemodel/batch.js';
 import Delta from '/ckeditor5/core/treemodel/delta/delta.js';
 import CKEditorError from '/ckeditor5/core/ckeditorerror.js';
