+++ tests/treemodel/delta/transform/delta.js
@@ -7,7 +7,11 @@
 
 'use strict';
 
-import transform from '/ckeditor5/core/treemodel/delta/transform/transform.js';
+import transformations from '/ckeditor5/core/treemodel/delta/basic-transformations.js';
+/*jshint unused: false*/
+
+import transform from '/ckeditor5/core/treemodel/delta/transform.js';
+
 import Position from '/ckeditor5/core/treemodel/position.js';
 import MoveOperation from '/ckeditor5/core/treemodel/operation/moveoperation.js';
 import Delta from '/ckeditor5/core/treemodel/delta/delta.js';
